std: wire view refactor
